gan_detect_iris/requirements.txt
scikit_image==0.18.1
scipy==1.6.1
matplotlib==3.3.4
numpy==1.20.1
dlib==19.21.1
Pillow==8.1.0
scikit-image==0.18.1
opencv-python==4.5.1.48
